support float

+++ src/Chapter2/Num.java
@@ -1,8 +1,8 @@
 package Chapter2;
 
 public class Num extends Token {
-    public final int value;
-    public Num(int value) {
+    public final float value;
+    public Num(float value) {
         super(Tag.NUM);
         this.value = value;
     }
